Move dd and help dialog into separate files
