+++ src/canvas/widgets.rs
@@ -1,7 +1,9 @@
 pub mod basic_table_arrows;
 pub mod cpu_basic;
 pub mod cpu_graph;
+pub mod dd_dialog;
 pub mod disk_table;
+pub mod help_dialog;
 pub mod mem_basic;
 pub mod mem_graph;
 pub mod network_basic;
@@ -12,7 +14,9 @@ pub mod temp_table;
 pub use basic_table_arrows::BasicTableArrows;
 pub use cpu_basic::CpuBasicWidget;
 pub use cpu_graph::CpuGraphWidget;
+pub use dd_dialog::KillDialog;
 pub use disk_table::DiskTableWidget;
+pub use help_dialog::HelpDialog;
 pub use mem_basic::MemBasicWidget;
 pub use mem_graph::MemGraphWidget;
 pub use network_basic::NetworkBasicWidget;
